Simplified event posting.
Previously, when an event gets posted to an endpoint, the
event name used to get appended to the given resid. This is
superfluous since the event object is itself being passed
(which I don't like yet, but since nothing else essentially
uses the event object, I'll grudgingly live with it for now).
Also removed passing on of query and headers for these
as they are unnecessary.