TERRIBLE_HACK: Support notion of application in ingress service.
I want to support the notion of an "application" that consists of
a number of services that work together. The application serves as
a namespace or mount point under which all activity about the application
happens. In essence, it serves as a URL interpretation point for the
other services none of which need to be exposed.

To satisfy this, I need to seamlessly support _codebase and serviceId/_proxy
as well, which are required for normal functionality. This is easiest to
add support for in ingress, BUT THAT MIXES RESPONSIBILITY HORRIBLY and hence
this is a TERRIBLE_HACK.

TODO: FIX THIS!